[chore] Separate MyPage and PointPage
