+++ src/App.jsx
@@ -5,7 +5,7 @@ import AnswerEndPage from "./pages/AnswerEndPage";
 import LoginPage from "./pages/LoginPage";
 import AnswerPage from "./pages/AnswerPage";
 import SignUpPage from "./pages/signup/SignUpPage";
-import MyPage from "./pages/MyPage";
+import MyPage from "./pages/mypage/MyPage";
 import ResultPage from "./pages/ResultPage";
 import AnswerStartPage from "./pages/AnswerStartPage";
 import FindPasswordPage from "./pages/FindPasswordPage";
@@ -14,6 +14,7 @@ import SurveyModalPage from "./pages/SurveyModalPage";
 import MySurveyPage from "./pages/mysurveypage/MySurveyPage";
 import MyAnsweredPage from "./pages/mysurveypage/MyAnsweredPage";
 import LandingPage from "./pages/LandingPage";
+import MyPagePoint from "./pages/mypage/MyPagePoint";
 
 function App() {
   return (
@@ -25,6 +26,7 @@ function App() {
       <Route path="/answer" element={<AnswerPage />} />
       <Route path="/signup" element={<SignUpPage />} />
       <Route path="/mypage" element={<MyPage />} />
+      <Route path="/mypage/point" element={<MyPagePoint />} />
       <Route path="/result" element={<ResultPage />} />
       <Route path="/answerstart" element={<AnswerStartPage />} />
       <Route path="/find-id" element={<FindIdPage />} />
